Change way of fetching single key to fix extendability
Hi there

This change makes it possible to fetch single settings by explicitly providing the column by which it should be searched for, instead of relying on the key column being the primary id.

This change is necessary to extend the model with a default ID column for it be referenceable by e.x. the activity log package from spatie. simple change, huge effect.

Thx
Best
Dimitri